Re-disable logevent compression (#403)
This was explicitly disabled in v5.12.1 and v5.13.2. Then it was moved
to an SDK flag without fixing the underlying problem. We should not rely
on an sdk flag for functionality that is definitively incorrect. Lets
remove this flag check entirely until the underlying issue is fixed. We
will likely re-release using a different flag