Optimize exposure cleaning in evaluator (#470)
I noticed some CPU time was being spent in _cleanExposures for gate
evaluation. It seems like this is called a bit excessively currently
(fully re-dedupes after each rule evaluation), so just optimizing the
happy path a bit. Ideally the exposures array would probably persist the
map for incremental deduping if we want to optimize cases with many
secondary exposures, but that's likely not as common.

- Avoids some memory allocation (new dedupe map, three new copies of the
array after .filter() and .map()) in the common case of no exposures to
clean
- Consolidates the three lambdas (filter, map, filter) into a single
filter to avoid making extra copies of the array

---------

Co-authored-by: tore-statsig <74584483+tore-statsig@users.noreply.github.com>